Fix requirements for py2app
There look to be a few outstanding bugs in py2app and its dependencies. I found the combination of dependencies that work here: http://aodj.ghost.io/isolation-in-practice-part-2/.

The process I am using to build is:

1) create a virtualenv and install mimic, it's dependencies, and the dependencies from the py2app-requires.txt file.
2) run python setup-mac.py py2app.

This is yieled an application, but it is crashing. I'm pretty sure the issues are related to either my start-app.py script being faulty or py2app not picking up a dependency.